refine ppdet file name (#2527)
* refine ppdet file name

* unify bbox_utils
